Sandbox database changes, to implement statement report.
The transactions table becomes a LongIntId, and a new table
now holds the fresh (= not yet included in a statement) transactions
until new statements get generated.  After that, the table of
fresh transactions get emptied.

Making the transaction table a LongIntId helps to keep
references to it from within the table of fresh transactions.