Activating withdrawals from the Sandbox.
At this point, the Sandbox offers a /taler endpoint that
creates a withdraw operation and returns (only) the corresponding
taler://-URI that should then be passed to the CLI wallet.

For simplicity, the Sandbox expects to have three canonical
bank accounts: exchange, customer, merchant, and refuses to
continue whenever one is not found.

Lastly, Sandbox now expects the hostname to be specified in
a environment variable; it should be noted that another way
to retrieve the hostname is already under development, but it
is not finished yet.