Amount representation.
Defer the conversion of amount strings into
BigDecimal until the point where they act as
numeric operands.

This saves resources because in several cases
the amount strings do not partecipate in any
calculation.  For example, an error might occur
before the calculation, or the calculation is
not carried at all by the function that gets
the amount string from the network.