GET /accounts DB logic, fix POST /transfer idempotence.
Along this change, the helper that strips IBAN payto URI
got changed to return null on invalid input.  The reason
is (1) a more convenient error handling by the caller: Elvis
operator instead of a try-catch block and (2) the impossibility
of 'util' to throw LibeufinBankException to drive Ktor to any
wanted error response.