Fixing mock payment initiator.
It needed a well-formed receiver BIC and name and a timestamp
not older than 3 months.  It's been observed however that the
bank test platform responded with EBICS_OK but did not show the
uploaded files in the Web UI.

Uploading the same content manually in the Web UI resulted instead
in a pain.002 report informing that the file was not valid.