nexus fetch parser
Parsing incoming and outgoing payments in two different
functions.  That is because two different camt.054 formats
("Detailavisierung" for incoming, "Lastschriftavisierung"
for outgoing) bring them.

Using MsgId, matching what was specified in the pain.001
and as well mentioned in the camt.054 "Lastschriftavisierung",
as the UID for the outgoing payments.  This choice helps also
to match the camt.054 entry with its related payment initiation
for reconciliation.